+++ CHANGELOG.md
@@ -6,6 +6,19 @@ Using for transmitting variables from External Tasks to another external workers
 
 ## [Unreleased]
 
+### Planned
+- Set non-blocking mode of receiving messages for Rabbit MQ connection
+- Set QoS parameter for channel: one message per one loop for Rabbit MQ connection
+- Set manual acknowledge for received message for Rabbit MQ connection
+
+## [0.2] - 2019-12-17
+
+### Added
+
+- Added basic auth for API Camunda BPM in config
+- Added validate message headers in connector-out 
+- Added validate Camunda external params in connector-in
+
 ## [0.1] - 2019-12-13
 
 ### Added
@@ -15,5 +28,6 @@ Using for transmitting variables from External Tasks to another external workers
 - Worker for test connector
 - README and CHANGELOG
 
-[unreleased]: https://gitlab.com/quancy-core/bpm-connector/-/tags/v0.1
+[unreleased]: https://gitlab.com/quancy-core/bpm-connector/-/tags/v0.2
+[0.2]: https://gitlab.com/quancy-core/bpm-connector/-/tags/v0.2
 [0.1]: https://gitlab.com/quancy-core/bpm-connector/-/tags/v0.1